ct: do not allow empty host mac address for CT #PSBM-152520
We should use predefined mac address for CT on the Host machine
to avoid reconfigure Bridge on the Host.
Predefined mac address should be less than mac address of
Bridge on the host. VZ9 use prefix of mac address 'FE3772'
for this purposes.
Empty host mac address means that libvzctl scripts should
create it but without adjusting, this brings for us
randomly generated mac addresses and effects proper work
of network on the Host.

Signed-off-by: Aleksandr Leskin <aleksandr.leskin@virtuozzo.com>