Omit IOTUNE_GROUP_NAME flag in case zero IoTune limit setting #PSBM-151497
Libvirt function qemuDomainSetGroupBlockIoTune in case of IoTune setting
expects that if iolimit is set to zero, the blk group name will not be
specified, thus libvirt interprets a null io value as a way to remove
the blk disk group.
Otherwise, libvirt throws an error because creating a blk group
with zero parameters makes no sense.
We omit specifying the flag VIR_DOMAIN_BLOCK_IOTUNE_GROUP_NAME
and group name in case of  iolimit zero setting.

Signed-off-by: Fima Shevrin <efim.shevrin@virtuozzo.com>