Add mac and hostInterface name setting forcibly #PSBM-155705
In case of creating a VM via prlctl, macaddress and hostInterface
name are generated during
VM registration within TaskRegisterVM. Here we add default behavior
for virsh define similar to prlctl create. If dumpxml of a VM contains
a network interface, but this interface does not have a hostInterface name,
then this device will not be validated by the rng schema of the dispatcher
and the dispatcher will send to libvirt is an empty xml
and libvirt will be crushed with sigfault.

Signed-off-by: Fima Shevrin <efim.shevrin@virtuozzo.com>